refactor: move shapeRepApi methods to command classes
