+++ src/commands/force/org/shape/delete.ts
@@ -7,11 +7,27 @@
 
 import { EOL } from 'os';
 import { flags, FlagsConfig, SfdxCommand } from '@salesforce/command';
-import { SfdxError, Messages } from '@salesforce/core';
-import { ShapeRepresentationApi } from '../../../../shared/shapeRepApi';
+import { SfdxError, Messages, Connection } from '@salesforce/core';
+import { RecordResult } from 'jsforce';
+import { isShapeEnabled } from '../../../../shared/orgShapeListUtils';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'shape.delete');
+
+interface ShapeRepresentation {
+  Id: string;
+  Status: string;
+  Edition: string;
+  Features: string;
+  Settings: string;
+  CreatedDate: string;
+  Description: string;
+}
+
+interface JsForceError extends Error {
+  errorCode: string;
+  fields: string[];
+}
 
 export interface OrgShapeDeleteResult {
   orgId: string;
@@ -27,6 +43,7 @@ export class OrgShapeDeleteCommand extends SfdxCommand {
       description: messages.getMessage('noPrompt'),
     }),
   };
+  private conn: Connection;
 
   public async run(): Promise<OrgShapeDeleteResult> {
     if (!this.flags.noprompt) {
@@ -34,11 +51,13 @@ export class OrgShapeDeleteCommand extends SfdxCommand {
         return;
       }
     }
-    const api = new ShapeRepresentationApi(this.org);
-    if (!(await api.isFeatureEnabled())) {
+
+    this.conn = this.org.getConnection();
+
+    if (!(await isShapeEnabled(this.conn))) {
       throw new SfdxError(messages.getMessage('noAccess', [this.org.getUsername()]));
     }
-    const deletedShapesIds = await api.deleteAll();
+    const deletedShapesIds = await this.deleteAll();
     if (deletedShapesIds.length === 0) {
       this.ux.log(messages.getMessage('noShapesHumanSuccess', [this.org.getOrgId()]));
     }
@@ -47,5 +66,39 @@ export class OrgShapeDeleteCommand extends SfdxCommand {
       orgId: this.org.getOrgId(),
       shapeIds: deletedShapesIds,
     };
+  }
+
+  /**
+   * Delete all ShapeRepresentation records for the shapeOrg.
+   *
+   * @return List of SR IDs that were deleted
+   */
+  private async deleteAll(): Promise<string[]> {
+    let shapeIds = [];
+    try {
+      const result = await this.conn.query<ShapeRepresentation>('SELECT Id FROM ShapeRepresentation');
+      shapeIds = result.records.map((shape) => shape.Id);
+    } catch (err) {
+      const JsForceErr = err as JsForceError;
+      if (JsForceErr.errorCode && JsForceErr.errorCode === 'INVALID_TYPE') {
+        // ShapeExportPref is not enabled, or user does not have CRUD access
+        throw SfdxError.wrap(messages.getMessage('delete_shape_command_no_access', shapeIds));
+      }
+      // non-access error
+      throw JsForceErr;
+    }
+
+    return Promise.all(
+      shapeIds.map(async (id) => {
+        try {
+          const delResult: RecordResult = await this.conn.sobject('ShapeRepresentation').delete(id);
+          if (delResult.success) {
+            return delResult.id;
+          }
+        } catch (err) {
+          return Promise.reject(err);
+        }
+      })
+    );
   }
 }
